🧪 test: ran lint for tests

+++ src/table/tests/interactions_test.rs
@@ -65,7 +65,7 @@ fn test_up_down_navigation_input() {
 #[test]
 fn test_max_pages() {
     let desc = "Test max page math calculations for tables:";
-    let test_cases = [(3 as i64, "55 / 20"), (2, "40 / 20"), (1, "5 / 13 ")];
+    let test_cases = [(3_i64, "55 / 20"), (2, "40 / 20"), (1, "5 / 13 ")];
     test_cases
         .into_iter()
         .zip(TableData::mock_pages())
@@ -78,7 +78,7 @@ fn test_max_pages() {
 fn test_req_offset() {
     let desc = "Test request offset calculation for tables:";
     let test_cases = [
-        (40 as i64, "55 items, page 3, limit 20"),
+        (40_i64, "55 items, page 3, limit 20"),
         (20, "40 items, page 2, limit 20"),
         (0, "5 items, page 1, limit 13"),
     ];
